requirements.txt
smbprotocol
paramiko
watchdog 
PyQt5
scp
bs4
pyodbc
psutil
python-dotenv